Añadir variantes de platos

+++ backend/app/Product/Domain/ValueObject/Stock.php
@@ -4,9 +4,9 @@ namespace App\Product\Domain\ValueObject;
 
 class Stock
 {
-    private int $value;
+    private float $value;
 
-    private function __construct(int $value)
+    private function __construct(float $value)
     {
         if ($value < 0) {
             throw new \InvalidArgumentException('El stock no puede ser negativo');
@@ -15,12 +15,12 @@ class Stock
         $this->value = $value;
     }
 
-    public static function create(int $value): self
+    public static function create(float $value): self
     {
         return new self($value);
     }
 
-    public function value(): int
+    public function value(): float
     {
         return $this->value;
     }
